--- main.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:daily_taskmate/HomeScreen.dart';
-
-void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(debugShowCheckedModeBanner: false, home: HomeScreen());
-  }
-}
